feat: move portfolio data fetch to info
